+++ index.js
@@ -1,5 +1,8 @@
 import Components from './src/index'
 import { ServerTable, ClientTable } from 'maestro-vue-tables-2'
+import VueNotifications from 'vue-notifications'
+import iziToast from 'izitoast'
+
 
 function callerComp (Vue, Comps) {
   for (let key in Comps) {
@@ -8,6 +11,15 @@ function callerComp (Vue, Comps) {
 
   Vue.use(ServerTable, {})
   Vue.use(ClientTable, {})
+
+  const toast = ({title, message, type, timeout, position = 'topRight'}) => iziToast[type]({title, message, timeout, position})
+  const options = {
+    success: toast,
+    error: toast,
+    info: toast,
+    warning: toast
+  }
+  Vue.use(VueNotifications, options)
 }
 
 function plugin (Vue) {