+++ src/_core/_utils/popoverMixins.js
@@ -23,31 +23,33 @@ export default {
       setTimeout(() => this.position(), 30)
     },
     position () {
+      let popover = this.$refs.popover
+      let trigger = this.$refs.trigger.children[0]
       this.$nextTick(() => {
-        let popover = this.$refs.popover
-        let trigger = this.$refs.trigger.children[0]
-        switch (this.placement) {
-          case 'top' :
-            this.left = trigger.offsetLeft - popover.offsetWidth / 2 + trigger.offsetWidth / 2
-            this.top = trigger.offsetTop - popover.offsetHeight
-            break
-          case 'left':
-            this.left = trigger.offsetLeft - popover.offsetWidth
-            this.top = trigger.offsetTop + trigger.offsetHeight / 2 - popover.offsetHeight / 2
-            break
-          case 'right':
-            this.left = trigger.offsetLeft + trigger.offsetWidth
-            this.top = trigger.offsetTop + trigger.offsetHeight / 2 - popover.offsetHeight / 2
-            break
-          case 'bottom':
-            this.left = trigger.offsetLeft - popover.offsetWidth / 2 + trigger.offsetWidth / 2
-            this.top = trigger.offsetTop + trigger.offsetHeight
-            break
-          default:
-            console.warn('Wrong placement prop')
+        if (trigger && popover) {
+          switch (this.placement) {
+            case 'top' :
+              this.left = trigger.offsetLeft - popover.offsetWidth / 2 + trigger.offsetWidth / 2
+              this.top = trigger.offsetTop - popover.offsetHeight
+              break
+            case 'left':
+              this.left = trigger.offsetLeft - popover.offsetWidth
+              this.top = trigger.offsetTop + trigger.offsetHeight / 2 - popover.offsetHeight / 2
+              break
+            case 'right':
+              this.left = trigger.offsetLeft + trigger.offsetWidth
+              this.top = trigger.offsetTop + trigger.offsetHeight / 2 - popover.offsetHeight / 2
+              break
+            case 'bottom':
+              this.left = trigger.offsetLeft - popover.offsetWidth / 2 + trigger.offsetWidth / 2
+              this.top = trigger.offsetTop + trigger.offsetHeight
+              break
+            default:
+              console.warn('Wrong placement prop')
+          }
+          popover.style.top = this.top + 'px'
+          popover.style.left = this.left + 'px'
         }
-        popover.style.top = this.top + 'px'
-        popover.style.left = this.left + 'px'
       })
     },
     toggle (e) {