+++ src/toast/index.js
@@ -1,0 +1,2 @@
+import comp from './Toast.vue'
+export default comp